Wizard/FSC: Deduplicate `useGetOscapCustomizationsQuery` request
The request would be called for every row. RTK query would have deduplicated the requests, but this is cleaner approach, with one subscription instead of N.
